Adding top-level documentation comments to every source file. Moved NoteGridViewModelModule into its own file.
