+++ app/src/main/java/com/kaelambda/note_grid/NoteGridApplication.kt
@@ -3,5 +3,8 @@ package com.kaelambda.note_grid
 import android.app.Application
 import dagger.hilt.android.HiltAndroidApp
 
+/**
+ * This is required for Hilt integration.
+ */
 @HiltAndroidApp
-class NoteGridApplication : Application() { }
+class NoteGridApplication : Application()
