+++ app/src/main/java/com/kaelambda/note_grid/screen/NoteGridViewModel.kt
@@ -1,24 +1,21 @@
 package com.kaelambda.note_grid.screen
 
-import android.content.Context
 import android.media.MediaPlayer
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
-import cn.sherlock.com.sun.media.sound.SoftSynthesizer
 import com.kaelambda.note_grid.audio.MidiFileWriter
 import com.kaelambda.note_grid.audio.MidiSoundController
 import com.kaelambda.note_grid.audio.SoundPoolController
-import dagger.Module
-import dagger.Provides
-import dagger.hilt.InstallIn
-import dagger.hilt.android.components.ViewModelComponent
 import dagger.hilt.android.lifecycle.HiltViewModel
-import dagger.hilt.android.qualifiers.ApplicationContext
-import java.io.File
 import javax.inject.Inject
-import javax.inject.Qualifier
 
+/**
+ * Our screen's ViewModel contains references to various controllers of business logic, a
+ * MediaPlayer, and functions for responding to events triggered by the UI.
+ *
+ * The 'useMidi' state variable is hoisted here so that it can be used in the ViewModel's logic.
+ */
 @HiltViewModel
 class NoteGridViewModel @Inject constructor() : ViewModel() {
     @Inject lateinit var soundPoolController: SoundPoolController
@@ -58,31 +55,5 @@ class NoteGridViewModel @Inject constructor() : ViewModel() {
     override fun onCleared() {
         super.onCleared()
         soundPoolController.release()
-    }
-}
-
-@Module
-@InstallIn(ViewModelComponent::class)
-object NoteGridViewModelModule {
-    @Qualifier
-    @Retention(AnnotationRetention.BINARY)
-    annotation class OutputFile
-
-    @OutputFile
-    @Provides
-    fun provideOutputFile(@ApplicationContext appContext: Context) : File {
-        val outputFile = File(appContext.filesDir, "exampleout.mid")
-        outputFile.createNewFile()
-        return outputFile
-    }
-
-    @Provides
-    fun provideMediaPlayer(): MediaPlayer {
-        return MediaPlayer()
-    }
-
-    @Provides
-    fun providesSynth(): SoftSynthesizer {
-        return SoftSynthesizer()
     }
 }
